normalizer to converter
