+++ include/SPTK/converter/filter_coefficient_normalization.h
