+++ include/SPTK/converter/generalized_cepstrum_gain_normalization.h
