+++ include/SPTK/converter/generalized_cepstrum_inverse_gain_normalization.h
