+++ src/main/mgcep.cc
@@ -55,7 +55,7 @@
 #include "SPTK/analyzer/mel_generalized_cepstral_analysis.h"
 #include "SPTK/converter/mel_cepstrum_to_mlsa_digital_filter_coefficients.h"
 #include "SPTK/converter/waveform_to_spectrum.h"
-#include "SPTK/normalizer/generalized_cepstrum_gain_normalization.h"
+#include "SPTK/converter/generalized_cepstrum_gain_normalization.h"
 #include "SPTK/utils/sptk_utils.h"
 
 namespace {
